+++ quiz-client/src/App.js
@@ -1,28 +1,13 @@
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
-import { useDispatch, useSelector } from "react-redux";
-import { TopicAPI } from "./api/topic-api";
-import { setTopicList } from "./store/topics/topics-slice";
-import { useEffect } from "react";
 import Home from "./pages/Home/Home";
 import Detail from "./pages/Detail/Detail";
 import Quiz from "./pages/Quiz/Quiz";
 import { SignIn } from "./pages/SignIn/SignIn";
 import { ProtectedPageBrowser } from "./pages/PageBrowser/PageBrowser";
+import PageNotFound from "./pages/404/PageNotFound";
 
 function App() {
-  const dispatch = useDispatch();
-  const user = useSelector((store) => store.authSlice.auth.user);
-
-  const fetchAll = async () => {
-    const topicList = await TopicAPI.fetchAllIncludeQuestionGroup();
-    dispatch(setTopicList(topicList));
-  };
-
-  useEffect(() => {
-    fetchAll().catch(console.error);
-  }, []);
-
   return (
     <div className="App">
       {/* <SideMenu /> */}
@@ -32,6 +17,7 @@ function App() {
           <Route path="/" element={<Home />} />
           <Route path="/detail/:questionGroupId" element={<Detail />} />
           <Route path="/start/:questionGroupId" element={<Quiz />} />
+          <Route path="*" element={<PageNotFound />} />
         </Route>
       </Routes>
     </div>
